[CONFIGURATION] * NEW: Switch to 3.3.90 version.
[WEBSITE MAKER]
* NEW: Update for 3.3.2 version.

[INTERPRETER]
* NEW: New API for getting information about a file.

[GB.DESKTOP]
* NEW: _Desktop.Time is a new property that returns the last X11 timestamp.
* NEW: _Desktop.ActivateWindow() is a new method that will be removed in a 
  next commit.
* NEW: DesktopWindow.Activate() is a new method to activate a window.

[GB.GTK]
* NEW: New GB_INFO constant to return the last X11 timestamp to other 
  components.

[GB.HTTPD]
* NEW: Reindent source files.

[GB.QT4]
* NEW: New GB_INFO constant to return the last X11 timestamp to other 
  components.

[GB.QT4.WEBKIT]
* BUG: Fix WebHitTest.Editable property declaration.


git-svn-id: svn://localhost/gambas/trunk@5221 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_archive.c
@@ -384,8 +384,8 @@ void ARCHIVE_stat(ARCHIVE *arch, const char *path, FILE_STAT *info)
   info->mtime = (int)buf.st_mtime;
   info->ctime = (int)buf.st_mtime;
   info->hidden = (*FILE_get_name(path) == '.');
-  info->uid = buf.st_uid;
-  info->gid = buf.st_gid;
+  info->uid = (int)buf.st_uid;
+  info->gid = (int)buf.st_gid;
 }
